requirements.txt
streamlit
yfinance
pandas
werkzeug
scikit-learn
streamlit-option-menu
requests
psycopg2-binary
extra-streamlit-components